increase frequency of backup age check
For my Postgres upgrade procedure, I'm triggering a backup manually
right before tearing down the old datadir. But the only way to trigger a
backup run manually is to reset the last_backup_timestamp. Right now,
it will take up to 5 minutes until the backup container notices that and
makes a new backup. This change reduces the wait time to max. 60
seconds.

This does not lead to more backups being produced. It's just about how
often the last_backup_timestamp is downloaded from Swift for checking if
a backup needs to be made.